chore(console): prepare and publish the design system (#3127)
- Publishes the `@wingconsole/design-system` package
- Exposes the tailwind plugins of both `@wingconsole/design-system` and `@wingconsole/ui` packages
