+++ apps/wing-console/console/design-system/package.json
@@ -2,19 +2,43 @@
   "name": "@wingconsole/design-system",
   "license": "SEE LICENSE IN LICENSE.md",
   "version": "0.0.0",
-  "private": true,
-  "exports": "./src/index.ts",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts",
+    "./tailwind-plugin.cjs": "./tailwind-plugin.cjs"
+  },
   "types": "./src/index.ts",
+  "publishConfig": {
+    "exports": "./dist/index.js",
+    "types": "./dist/index.d.ts",
+    "files": [
+      "dist",
+      "tailwind-plugin.cjs"
+    ]
+  },
   "scripts": {
+    "compile": "tsup",
     "tsc": "tsc --build",
     "eslint": "eslint --ext .js,.cjs,.ts,.cts,.mts,.tsx --no-error-on-unmatched-pattern . --fix",
-    "vitest": "vitest run --coverage --passWithNoTests"
+    "vitest": "vitest run --coverage --passWithNoTests",
+    "package": "bump-pack -b"
   },
-  "devDependencies": {
-    "@babel/core": "^7.22.5",
+  "dependencies": {
     "@headlessui/react": "^1.7.15",
     "@heroicons/react": "^2.0.18",
     "@popperjs/core": "^2.11.8",
+    "@tailwindcss/forms": "^0.5.3",
+    "classnames": "^2.3.2",
+    "lodash.uniq": "^4.5.0",
+    "nanoid": "^4.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-lottie-player": "^1.5.4",
+    "react-popper": "^2.3.0",
+    "tailwindcss": "^3.3.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.22.5",
     "@storybook/react": "^7.0.20",
     "@testing-library/react": "^14.0.0",
     "@types/lodash.uniq": "^4.5.7",
@@ -24,15 +48,10 @@
     "@vitest/coverage-c8": "^0.31.4",
     "@wingconsole/eslint-plugin": "file:../../tools/eslint-plugin",
     "@wingconsole/tsconfig": "file:../../tools/tsconfig",
-    "classnames": "^2.3.2",
+    "bump-pack": "file:../../../../tools/bump-pack",
     "eslint": "^8.42.0",
     "happy-dom": "^9.20.3",
-    "lodash.uniq": "^4.5.0",
-    "nanoid": "^4.0.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-lottie-player": "^1.5.4",
-    "react-popper": "^2.3.0",
+    "tsup": "^7.1.0",
     "typescript": "^5.1.3",
     "vite": "^4.3.9",
     "vitest": "^0.31.4",
