+++ apps/wing-console/console/design-system/project.json
@@ -1,23 +1,18 @@
 {
-  "name": "console-ui",
+  "name": "console-design-system",
   "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "implicitDependencies": [
-    "compiler",
-    "sdk",
-    "console-server",
-    "console-design-system"
-  ],
+  "implicitDependencies": [],
   "targets": {
     "build": {
       "options": {
         "command": "npm run compile",
-        "cwd": "apps/wing-console/console/ui"
+        "cwd": "apps/wing-console/console/design-system"
       }
     },
     "package": {
       "options": {
         "command": "npm run package",
-        "cwd": "apps/wing-console/console/ui"
+        "cwd": "apps/wing-console/console/design-system"
       }
     }
   }
