+++ apps/wing-console/console/ui/package.json
@@ -2,14 +2,17 @@
   "name": "@wingconsole/ui",
   "license": "SEE LICENSE IN LICENSE.md",
   "version": "0.0.0",
-  "type": "module",
-  "exports": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./tailwind-plugin.cjs": "./tailwind-plugin.cjs"
+  },
   "types": "./src/index.ts",
   "publishConfig": {
     "exports": "./dist/index.js",
     "types": "./dist/index.d.ts",
     "files": [
-      "dist"
+      "dist",
+      "tailwind-plugin.cjs"
     ]
   },
   "scripts": {
@@ -28,6 +31,7 @@
     "@trpc/client": "^10.30.0",
     "@trpc/react-query": "^10.30.0",
     "@trpc/server": "^10.30.0",
+    "@wingconsole/design-system": "file:../design-system",
     "classnames": "^2.3.2",
     "d3-selection": "^3.0.0",
     "d3-zoom": "^3.0.0",
@@ -52,7 +56,6 @@
     "@storybook/react": "^7.0.20",
     "@storybook/react-vite": "^7.0.20",
     "@storybook/testing-library": "^0.1.0",
-    "@tailwindcss/forms": "^0.5.3",
     "@types/cors": "^2.8.13",
     "@types/d3-selection": "^3.0.5",
     "@types/d3-zoom": "^3.0.3",
@@ -62,7 +65,6 @@
     "@types/react": "^18.2.12",
     "@types/react-dom": "^18.2.5",
     "@vitejs/plugin-react-swc": "^3.3.2",
-    "@wingconsole/design-system": "file:../design-system",
     "@wingconsole/eslint-plugin": "file:../../tools/eslint-plugin",
     "@wingconsole/server": "file:../server",
     "@wingconsole/tsconfig": "file:../../tools/tsconfig",
